WIP: Refactering.
